sentimini/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
defusedxml==0.4.1
dj-database-url==0.4.0
Django==1.9.7
django-allauth==0.25.2
django-bower==5.1.0
django-crispy-forms==1.6.0
django-nvd3==0.9.7
gunicorn==19.4.5
Jinja2==2.8
kombu==3.0.30
MarkupSafe==0.23
oauthlib==1.1.2
psycopg2==2.6.1
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.0.10
pytz==2016.4
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
Unidecode==0.4.19
whitenoise==2.0.6
